Add a development docker-compose
This will help me contribute to koel and ease testing.

How to test:
- docker-compose up -d --build
- docker-compose exec -it web bash
- In the container: php artisan koel:init
- It will fail at the "frontend stuff" stage. It is expected because
  Yarn is not installed in the container (and does not need to be).
Sources are supposed to be built on yout host machine.
- Connect to http://localhost
- You should see a Koel instance

See CONTRIBUTING.md for additional helpful commands and tips for
contributors.

Note: I spent hours trying to debug a SQLSTATE 2002 error. It turns out that
it was a cache issue -_- ... The command below fixed it somehow ???

  php artisan config:cache